Added remove product

+++ src/main/scala/model/Product.scala
@@ -1,10 +1,6 @@
 package model
 
-case class Product(
-    productId: Integer,
-    name: String,
-    price: Double
-) {
+case class Product(productId: Integer, name: String, price: BigDecimal) {
   require(productId > 0, "productId must be greater than 0")
 
   def isEmpty: Boolean = name.trim.isEmpty
